refactor: JSONLoader 네트워킹 로직 분리 및 DIP 적용
