+++ BoxOffice/Model/Fetcher.swift
@@ -1,10 +1,12 @@
 import Foundation
 
 struct Fetcher {
-     func fetchMovieInfo(movieCode: String) {
+    let networkManager: NetworkManagerProtocol = NetworkManager()
+    
+    func fetchMovieInfo(movieCode: String) {
         var movieInfoData = EndPoint.moviInfoPath
         movieInfoData.getDetailMovieInfoQueryItems(movieCode: movieCode)
-        JSONLoader().loadJSONFromURL(from: movieInfoData.url.absoluteString) { (result: Result<MovieInfoResponse, Error>) in
+        networkManager.fetchDetails(from: movieInfoData.url.absoluteString) { (result: Result<MovieInfoResponse, Error>) in
             switch result {
             case .success(let response):
                 dump(response)
@@ -14,10 +16,10 @@ struct Fetcher {
         }
     }
     
-     func fetchBoxOffice(targetDate: String) {
+    func fetchBoxOffice(targetDate: String) {
         var boxOfficeData = EndPoint.boxOfficePath
         boxOfficeData.getBoxOfficeQueryItems(targetDate: targetDate)
-        JSONLoader().loadJSONFromURL(from: boxOfficeData.url.absoluteString) { (result: Result<BoxOfficeResponse, Error>) in
+        networkManager.fetchDetails(from: boxOfficeData.url.absoluteString) { (result: Result<BoxOfficeResponse, Error>) in
             switch result {
             case .success(let boxOfficeResponse):
                 dump(boxOfficeResponse)
@@ -27,3 +29,4 @@ struct Fetcher {
         }
     }
 }
+
