+++ BoxOffice/Model/JSONLoader.swift
@@ -10,25 +10,4 @@ struct JSONLoader {
         let decoder = JSONDecoder()
         return  try decoder.decode(T.self, from: data)
     }
-    
-    func loadJSONFromURL<T: Codable>(from urlString: String, completion: @escaping (Result<T, Error>) -> Void) {
-        guard let url = URL(string: urlString) else {
-            completion(.failure(NetworkError.urlError))
-            return
-        }
-        let task = URLSession.shared.dataTask(with: url) { data, response, error in
-            guard let data = data, error == nil else {
-                completion(.failure(NetworkError.dataLoadingError))
-                return
-            }
-            let decoder = JSONDecoder()
-            do {
-                let decodedData = try decoder.decode(T.self, from: data)
-                completion(.success(decodedData))
-            } catch {
-                completion(.failure(NetworkError.decodingError))
-            }
-        }
-        task.resume()
-    }
 }
